Improve Autostart INF file parsing
This will now parse both autostart and resolution lines from the
in-memory INF file, in a way that can be easily later switched to
on-disk INF files. This is in preparation for joining VDI resolution
code and adding user overridable resolution settings.

Also, different TOS autostart IDs are now used for VT52 (TOS/TTP) and
GEM applications.