Fix problem with first parameter of f_out() being NULL
Recent versions of GCC complain that f_out() and thus fprintf is called
with the first parameter being zero. And indeed the WinUAE code expects
to be able to do this, e.g. when ShowEA() is called with 0 as first
parameter. Thus let's turn f_out() into a real function that checks
the first parameter, just like its pendant is doing it.