Remove support for obsolete VBCC/vlink pre-release
Not sure whether this issue was in any of vlink releases, or just in
the private pre-release I got.  It may also have been bug in Hatari's
program symbol table type recognization as nowadays my older VBCC
built programs seem to be recognized as a.out.

Improved also few messages show to user.