Ignore keyboard repeat events on SDL2
In SDL1, keyboard repeat was disabled by default, which is what we want
since we handle keyboard repeat on the ST side on our own. In SDL2,
keyboard repeat is always enabled by default, so we continuously get
KEYDOWN events until the key is released. This has some ugly effects
with certain shortcut keys, so we have to explicitely ignore keyboard
repeat events with SDL2.