Rename bOverlayState to nOverlayState and add /* fall through */ comment
The bOverlayState variable is not a boolean, so this variable should not
use the conventions for boolean variables and thus not start with a "b".

Also add a missing "/* fall through */" comment here to be able to compile
with "-Wimplicit-fallthrough=1" without warnings here. (The comment right
before the switch statement says that all states except none need to be
updated on the screen, so we indeed need the fall-through instead of a
break here)