Fix Coverity warning about suspicious implicit sign extension
sdlscrn->format->BytesPerPixel is Uint16. This is promoted to
"int" when multiplied with scrwidth (which is "int", too). This
could then be sign-extended when used as third argument of
memcpy(). So make sure that the result of the multiplication
is unsigned by using a temporary "unsigned int" variable for
the BytesPerPixel value.