Fix broken CPU dialog with old UAE CPU core
The dialog uses a different amount of entries in the FPU box for the
old UAE CPU core since the FPU_JIT flag is only used with the new
WinUAE core. So the entries in cpudlg[] do not match the index macros
like DLGCPU_PREFETCH, which is causing some wrong effects when the
dialog is shown. Fix it by removing the FPU_JIT stuff completely. We
can add it again later in case we ever support JIT with Hatari.