Add the possibility to run Hatari with "--tos none"
We really need a way to run regression tests automatically. The main
problem is our dependency to external TOS ROM images for this. But
since we've got code to handle many GEMDOS calls already, it should
be possible to fake enough of TOS to run our simple regression tests
already without an external TOS ROM image. So let's add a mode to run
Hatari with "--tos none", in which it loads, relocates and runs a simple
PRG file using our GEMDOS emulation instead of a real TOS. We also fake
a simple "ROM" image for this so that we do not have to change other
code in Hatari all over the place and the basic way of operation stays
the same (i.e. the instruction execution starts with the ROM image).