Move the check-bashism test into a separate script.
Using "find" to get the list with the shell scripts was a rather bad
idea since a user might have more scripts in the directory. Also
using the file name of scripts for creating the different test names
was rather cumbersome. So move the bashism test into a separate
script file and use "hg files" to get a list of shell scripts in
the repositry instead of using "find".
Thanks to Eero for the idea and the inital version of the script.