Do not run the checkbashisms test automatically
First, there is the problem that "hg files" does not work with older
versions of Mercurial. Easy to fix by using "hg locate" instead. But
then I also had to discover that there are older versions of
"checkbashisms" available which claim that "export foo=bar should be
foo=bar; export foo" - though this is valid POSIX shell scripting.
Thus the automatic test fails with these versions of checkbashisms.
Since this test is apparently very fragile, and Eero also complained
that it is way too slow on older computers, it seems best to disable
it by default and only run it manually from time to time on systems
where we know that we've got a proper installation with hg and a new
version of checkbashisms.