Add experimental NCR5380 SCSI support for Falcon mode
Code is based on the NCR5380 SCSI emulation code from WinUAE (from the
file scsi.cpp there), but it has been changed quite a bit to work with
the SCSI command emulation in hdc.c from Hatari. It's enough to get sector
read support with the demo version of HDDRIVER already. DMA write support
is not implemented yet, so the code is still disabled by default (set the
WITH_NCR5380 macro to 1 to enable it).