More TOS 4.x patches for 68040 and 68060
These patches are a port of the corresponding patches that ARAnyM applies when
booting TOS 4.04. In a nutshell: the early MMU initialization (setting up the
transparent translation) at reset & coldboot is replaced by the corresponding
040 specific code. Several places where the cache for 030 is flushed are
replaced by cinva/cpusha. Some places where only the data cache is invalidated
are nopped out (not entirely sure about those, but works in ARAnyM and seems
to also work in Hatari, at least the current attempt of setting bit 11 in the
cacr won't have any effect on 040; most of these places seem to be in the
VDI). The _CPU cookie is set up correctly.
In the larger part that is put at the end of the ROM, there are some places
that are apparently not used (anymore), but they have been kept for simpli-
city because otherwise some absolute addresses would change.