Fix 24-bit address bus handling during memory_init()
During reset, the TOS image is loaded first, and m68k CPU reset is done
afterwards. This causes a problem with memory_init(), which is called
from the TOS loading function and uses currprefs.address_space_24, a
variable which we only update during the CPU reset, i.e. the
currprefs.address_space_24 variable still contains the old value in
case the user changed the setting in the GUI.
Fix it by using the up-to-date ConfigureParams.System.bAddressSpace24
instead.