Support the FASTLOAD flag in the program loading function of the cartridge
We should clear the whole heap if the FASTLOAD flag is not set. Some old
programs expect this - this fixes the graphical glitches in the endpart
scroller of the "Pacemaker" STE demo, for example.

While we're at it, also remove the cart_mus.x32 data. The assembler
instruction for including it is not portable (e.g. "incbin" vs. "ibytes")
and the easter-egg simply has had its days.