Linux Loader: add m68k Linux 030 issue workaround patch
This was there already in 4.x kernels, and is still there with v5.2.

Maybe some of the other issues are also due to kernel missing support
for ignoring certain exceptions on 030 (and for some reason those
exceptions being generated more often with Hatari)?

Adding patch more to inform about the issue, than the issue that patch
fixes being important.