Handle GEMDOS HD emulation without code in the cartridge
The assembler code in the cartridge for the GEMDOS HD emulation is a
constant source of trouble. Currently there is a problem that it
crashes in TT emulation mode when an interrupt from the 2nd MFP
happens at the wrong point in time and the stack supplied by TOS
is too small for handling both, the GEMDOS HD code in the cartridge
and the interrupt routine.

Instead of continuing the kludgy and hard-to-maintain assembler code
in the cartridge, let's rather act directly when we see the exception
in the Hatari C code. The only trouble is the Pexec() GEMDOS call
which needs some extra handling: We need to allocate memory on the ST
side before loading the program code, so we have to run some ST code
in between. Let's use the old GEMDOS_OPCODE illegal opcode now to jump
back to the emulator code for this.