Fix "[ffmpeg/video] Invalid PNG signature 0x89504E470D0A1A"
The attempt to align PNG chunk sizes with 16-bit word boundaries causes
problems when replaying the resulting AVI file using mpv:

	Error while decoding frame!
	[ffmpeg/video] png: Invalid PNG signature 0x89504E470D0A1A.

Removing the offending alignment code solves this problem.

Additionally, but unrelated to this change, there is an issue with
audio and video synchronisation, that is also problematic with the
BMP coder.