Add bus error tests with non-standard CPU modes
We recently had a regression that bus errors were not working anymore
with MMU enabled. We did not notice it immediately since we do not run
any bus error tests in these CPU modes yet. Thus add some more tests
with non-default CPU modes to avoid that this breaks so easily again.