Fix bugs in Dsetpath() and Dgetpath()
Dgetpath() has to return an empty string if the current directory is the
root directory of a drive.
And Dsetpath("") should simply do nothing but returning 0 for success
(and while we're at it, check for valid ST memory in that function, too).

Thanks to Roger Burrows for reporting the bugs!