Create file names with a dedicated function instead of using sprintf + PATHSEP
Using PATHSEP with sprintf and friends only works on systems that have a
similar path handling to unix-like operating systems. It does not work
on Amiga OS, for example, and due to this we already have some ugly #ifdefs
in our code base. We already have File_MakePath() to create file names with
a base directory, but it is not usable in these spots yet since that function
return a freshly allocated buffer. Thus introduce a new function called
File_MakePathBuf() now that can be used in these spots instead, so that we
can finally get rid of the ugly #ifdefs in the code.