Fix problem with bogus command bytes / HuSHI version 2.16
Martin van der Koelen reported that there is an issue with HuSHI-junior
v.2.16 - it sends apparently superfluous command bytes after the command
packet is already finished. Let's do not reset the byteCount until a new
command is started when the A1 pin is pushed to 0. Thanks to Martin for
the bug report and the original version of this patch.