Fix fullscreen test to work with the latest changes to the video code
The fullscreen routine from Flix does not delay the VBLANK for some more
lines, so there are only 274 visible lines on the screen - and the last
lines should be black now. The new behavior is different between ST and
STE mode, so we need to reference pictures now.
Also add --drive-led off to the run script to make sure that the drive
LED can never spoil the screenshot here.