Rename Ide_MmioIsAvailable() to Ide_IsAvailable() and add secondary disk there
We are using the function in more spots than to only check for the
availability of the MMIO region, so use a more general name here.

Also add ConfigureParams.Ide[1].bUseDevice to the check. While it is not
possible to boot from a secondary drive, it seems like some drivers can
still access it without a primary drive.