Make sure to not read samples from beyond the end of STRam in the crossbar
The dem EKO-Systems temporarily sets the DMA start address higher than
the DMA end address. This caused the crossbar code to read beyond
the STRam array, causing very bad sound in the "racing scene" of the
demo, or even crashes of the emulator. Thus make sure to limit the accesses
to the STRam array only.