TOS boot tester: override all relevant Hatari config options
This includes setting many options to values that are already Hatari
built-in defaults.

Alternative to this would be to point $HOME to the test directory,
create ".config/hatari/" directory structure under it, and write the
override config file there (as "hatari.cfg") without options that are
Hatari built-in defaults.

(This sets also HATARI_TEST, to disable global Hatari config use.)