Fix another possible crash in the IDE code
When Hatari has been started in non-Falcon mode, or in Falcon mode
with fast-boot enabled, and no IDE drive has been configured, the
emulator crashes when switching to Falcon mode in the GUI (or when
disabling the fast-boot option in case it was already in Falcon mode).
Add another check to correctly re-initialize the IDE subsystem in
this case.
Thanks to the Emphii and czietz for the bug report at atari-forum.com!