Falcon compatibility: movem fix impact - part 2, other updates
Went through rest of Falcon programs that are noted to work either
with or without MMU (not both), and added more info for ones that
still do with the movem fix.

Weirdly there are also couple of Falcon programs that work only with
EmuTOS, not TOS4.

I noticed only now that Laurent's Racer games were missing from this
list, so those are added also.

Added missing ':' separators between demo and group names.