Python UI: fix/improve file selector titles
Fix: specify for Gtk.FileChooserButton() that label is to be used as
file selector title.

Allow title to be specified also FselEntry, not just for file
selectors gotten from FselAndEjectFactory.

Use labels in front of options (in settings dialogs) also as their
file selector titles, and have all of them ending with ':' for
consistency.