Use a static array for the ide state instead of allocating it dynamically
The dynamic allocation of the IDEState keeps causing trouble when
switching between configurations - there is currently yet another
crash when saving a memory snapshot in Falcon mode and then
restoring it from the command line without specifying the machine
type.
Let's stop playing whack-a-mole with this problem and use a static
array for the IDEState instead.