Update the Cirrus-CI (fix FreeBSD, remove VS job)
Both, FreeBSD and the VisualStudio job, are currently failing in the
gitlab-CI. FreeBSD needs an update to a newer version. But the VS job
fails for unknown reason during the installation step, and  I cannot
get it working anymore. Let's remove it again, this container really
seems not stable enough for a reliable CI.