Use history for determining suitable disassembly default address
If disassembly starts from program counter (PC), user lacks context on
how the code ended in current position.  With this, when history is
enabled in debugger, disassembly is started from suitable address
(slightly) before PC.

Because trying to disassemble instructions from "random" address
before PC will often output incorrect disassembly, this scans
execution history for a valid instruction address that's close enough
to PC.

History addresses are valid until overwritten with differently aligned
instructions, which is extremely unlike to happen, so this should be
robust enough to be used always when history is enabled.