Fix: MIDI IRQ needs to be re-enabled on re-init
Midi_UnInit() removes the IRQ, but Midi_Init() does not set it,
only Midi_Reset() does, which was called only on emulation resets.

Midi_Reset() cannot be called from Midi_Init() itself because that
gets called also before emulation and interrupts are running, so it
needs to be called in addition to Midi_Init() when there's no
associated reset.

From Nicolas: Variables reset in Midi_Reset() in addition to starting
the timer is same as reseting any bytes that were currently in transit
(ie we didn't receive all 8 bits yet), which should be acceptable when
midi device is changed on the fly.

(Add also few TODOs for PortMidi related variables that would be good
to reset on MIDI device change.)