Update mode page 0x3f to return all available pages in the order requested by the SCSI specification (patch by Uwe Seimet) Note from Uwe : This patch updates mode page 0x3f to return all available pages in the order requested by the SCSI specification (i.e. page 0 last).
IMO the implementation of mode page 0x04 is wrong. It is definitely incomplete
because it ignores some CDB data. In particular it does not check the DBD bit:

A disable block descriptors (DBD) bit of zero indicates that the target
may return zero or more block descriptors in the returned MODE SENSE data
(see 8.3.3), at the target's discretion.  A DBD bit of one specifies that
the target shall not return any block descriptors in the returned MODE
SENSE data.