Fix the checks for the end of the video RAM
fvram_line does not necessarily point into the STRam[] area, e.g. when
rendering monochrome mode on ST/STE (it points to the FrameBuffer.pSTScreen
copy there instead), so we have to look at the video base address instead
and count whether we've hit STRamEnd during the rendering.

Thanks to Christian Zietz for spotting and analyzing the problem!