Screen width must be a multiple of 16 in VDI resolution modes
Bitplanes are interleaved in chunks of 16 pixels, so it does not make
sense to allow alignments less than 16 here. This fixes a bug reported by
GoKMasE here: https://www.atari-forum.com/viewtopic.php?p=435522#p435522
where the right 8 pixels of the screen are not right when running Hatari
with:

 hatari --machine tt --vdi on --vdi-width 1280 --vdi-height 960 --vdi-planes 4

Hatari wrongly sets the screen size to 904x672 in this case - and 904
is not divisable by 16, so the final pixels of each line get messed up.

Fixes: 918580938908 ("Allow VDI mode to emulate devices with smaller screens")