Fix possible crash in VDI mode when the ST program sets a bad video base
The ST program might change the screen address between the point in time
where we set the VideoBase variable and the point in time we're doing
the conversion in screenConvert.c, so the code in that file must not
use Video_GetScreenBaseAddr() to get the base address, but has to use
the address that we recorded earlier. This fixes a possible crash in
VDI mode with the tests/mem_end/scr_end.prg test program.