Reduce our dependencies to screen.h
screen.h requires the SDL headers, so it's bad to include it
in code that does not really depend on the SDL library, thus
drop the inclusion of screen.h wherever it is possible.
However, video.h currently depends on screen.h due to the
HBL_PALETTE_LINES and HBL_PALETTE_MASKS macros. Fix this
dependency by moving the HBLPalettes[] and HBLPaletteMasks[]
arrays from video.c to screen.c instead.