Replace strlcpy() with a better implementation
strlcpy() has issues with untrusted source strings, since it
is always trying to find out the length of source strings. So
for emulators like Hatari this could be dangerous when being
used on strings that originate from the Atari program side.
Let's use a better string copy function instead, similar to
strscpy() in the Linux kernel (see following URL for details:
https://lkml.org/lkml/2022/8/21/271 ). This also decreases
our dependency on the SDL headers since we were using
SDL_strlcpy() on systems that do not have a strlcpy() function
in their libc.