Rename SDL_UpdateRect() into Screen_UpdateRect()
The SDL_UpdateRect() function was a hack that we introduced
when adding support for the SDL 2.0. Since we dropped support
for SDL 1.2 a while ago, keeping the old name is not required
anymore. Additionally, the SDL_UpdateRect() prototype also
conflicts with some compatibility functions from Emscripten,
so let's use a better name for this function now instead.