gitlab-ci: Remove the "build-normal" job
gitlab.com recently introduced limits for the monthly CI minutes,
so we have to be more conscious about the jobs that we are running.
Almost everything that is handled by the "build-normal" job is
already tested in other pipelines, so we can certainly drop this
job to save some CI minutes. While we're at it, turn "build-alt"
into a job that really tests for a minimalist configuration instead.