Fix joy button 2 space key emulation
Earlier it worked in rapid autofire mode making it impossible
to use for doing just a single space key press:
	https://www.atari-forum.com/viewtopic.php?p=444078

This commit changes it to work like normal key presses.

Simulated space key is released when the button is released, which
among other things meaning also that key repeat works as expected.