Show PortMidi device prefix matches also in GUI
Midi functionality already did prefix matching on devices and opened
them, but it did not update the config value to full MIDI device name.
This was to avoid breaking prefix matching for devices which names
(e.g. port) changes frequently.

With this change, same matching is done for names shown in GUI, to
avoid confusing users.  Downside of this is that visiting devices
dialog WILL replace the prefix in config setting with a full device
name.  I.e. saving settings afterwards, will override the prefix
in config file.