Reduces the differences between Hatari and Previous
Setting CMAKE_OSX_ARCHITECTURES and CMAKE_OSX_DEPLOYMENT_TARGET must be set
before the call to project(). Otherwise it won’t be recognised. Therefore
building Hatari as universal binary was broken.
Also add automatic addition of SDL2.framework and png.framework to the bundle.
This way you get a runnable application out of the box (else you would have to
copy the frameworks to the bundle manually before to make it work). This is
necessary since Xcode 15. Also remove some unused stuff.

Thanks to Andreas Grabher for the patch!