Use ST palette for PNG screenshots if possible (patch by Brad Smith) Notes from Brad Smith (https://github.com/hatari/hatari/pull/30) : This allows PNG screenshots to use the ST palette where possible.
A palettized image has a few advantages. The file size is slightly smaller.
Any graphics program that can view the palette can now be used to see what
the ST's 16-colour palette looked like at the time of screenshot. It makes
palette-based editing of the image easier (e.g. changing colours).
Conversion into an ST-ready image format is also more straightforward.

This automatically falls back to a full RGB image if the image contains any
colours outside the current ST palette. This means that modes with more colours,
or mid-screen palette changes will still produce valid screenshots as full RGB.

This works for low res and med res video modes.

NOTE 2 : As the Hatari UI also uses its own colors, you need to check "Crop Statusbar"
in the "Hatari Screen" menu to ensure the status bar is not included in the
screenshot. If not, it will use more colors than the emulated screen and resulting
PNG will be saved as RGB image and not as palettized image.