NEO direct video memory dump option for screen snapshot (patch by Brad Smith) This needs to be enabled via Config.bNEOScreenSnapShot=TRUE (no UI added for it)
Notes from Brad Smith (https://github.com/hatari/hatari/pull/31 for more details) :
This creates a NEO file format screenshot, out of the data in FrameBuffer (screen.c).
This works for ST low, medium, and high resolutions. The palette is taken
from the centre line of the FrameBuffer for Low and Medium resolutions.
For High resolution don't seem to store per-line palette, but I instead extracted
the GenConvert palette for it.