DSP fix: JMP can form a long interrupt instruction
Motorola documentation explains that JSR forms a long interrupt
instruction, but JMP works also.

Extract from Motorola doc
A long interrupt is formed if one of the interrupt instructions fetched is a JSR instruction.
The PC is immediately released, the SR and the PC are saved in the stack, and the jump instruction
controls where the next instruction is fetched from. While either an unconditional jump or
conditional jump can be used to form a long interrupt, they do not store the PC on the stack;
therefore, there is no return path.

This fix allows VOXXX.PRG to work