Reset natfeat SCSI subsystem when doing a soft-reset from the Atari side
When you reset the Atari within the emulation (by calling os_start in SYSHDR)
SCSI Driver handles in nf_scsidrv.c may still be open. With a real Atari
there is no need to close a handle before a reset, because the HW reset
implicitly clears the handles.
With Hatari it's different, because the external handle in nf_scsidrv.c
remains open after a reset triggered by an application. As a result you
run out of external handles. This happens, for instance, when you partition
a native Linux device with HDDRUTIL and then let HDDRUTIL execute a reset.

To fix the issue, call the nf_scsidrv_reset() when the CPU executes
the RESET instruction which should happen everytime when the code jumps
back to the beginning of the TOS ROM.

Thanks to Uwe Seimet for this patch!