Change the defaul mapping of the SDLK_MINUS key
It's used for 0x35 in more languages than it is used for 0x0C, so use
the default for the majority here. That means that we now also have to
provide a function for US mapping since the default mapping now differs
for this country.
Thanks to Brad Smith for the idea/original patch!