Fix: Wrong MODE SENSE data length for page code 0x3f
The mode data length field when all pages are returned (page code
0x3f) is wrong. Instead of 44 it has to be 43. The SPC-5 specification
says: "The mode data length field indicates the number of bytes that
follow in the mode parameter list". "Follow" is the important word
here, because usually fields that contain a byte count do not count
themselves. You also see that something is wrong with the current
length field when running the SCSI Driver test suite.

Without this patch it says:
	Page 0 (current, not savable)

	Reading all mode pages with MODE SENSE (10)
	MODE SENSE (10) is not supported by device

=> empty line indicates that something is wrong.

With the attached patch it says:
	Page 0 (current, not savable)
	0e:00:08:00:00:00:08:00:00:02:00:00:00:00:00
	Reading all mode pages with MODE SENSE (10)
	MODE SENSE (10) is not supported by device

=> 0e:00:08:00:00:00:08:00:00:02:00:00:00:00:00 matches the page 0
   data returned by Hatari.

Btw. This data is not really correct, because it returns a hard-coded
sector size of 512 bytes even when a different physical sector size is
configured. But as long as this page is only there to please HDX it
does not matter.