Disallow 68881/68882 FPU in 040/060 mode
The real 68040/68060 CPU cannot work together with a 68881/68882 FPU,
so we should not offer this possibility in Hatari. Let's emit a warning
and fall back to the internal FPU in case the user got this configuration
wrong. While we're at it, let's also fall back to 68882 FPU in case the
user selected an internal FPU for the 68020/030 (instead of disabling
it completely).