Create a dedicated header file for endianness swapping functions
The swapping functions in the SDL3 changed their names again, so we
need to rework the related code. But instead of working around that
problem with SDL-version specific ifdefs, let's rather provide our
own header with endianness swapping functions to become more independent
from SDL in various files.

The basic code is already available in cpu/sysdeps.h and cpu/maccess.h,
and since this is Hatari-specific code that does not need to be sync'ed
with WinUAE anymore, let's simply consolidate it in includes/endianswap.h
now.