Events: support CLI bool & integer option disabling
Initial event variable value needs to be invalid for the options, so
that disabling an option is also detected as a change. This requires
changing bool variables to ints so they can store 3 values
(unset/off/on).

With this one can e.g. use fast-forward only during boot:
	--fast-forward on --fast-forward prg:off

Flag string options do not need to change because those can be
disabled with the "none" value:
	--trace os_base --trace prg:none

Note: All values are set and logged now regardless of previous value,
because logging only their changes was confusing (user: didn't I set
it, or did it not work?).