Show cursor when emulation is paused in full screen mode
Commit 149e6b18d4a96da65a1c6ba043d862b43ba99f10 removed the extra
code that was enabling the mouse cursor before showing a dialog,
assuming that Main_PauseEmulation() would always enable it. However,
this did not happen in fullscreen mode, so there was no mouse cursor
showing up anymore for using the dialogs. To fix the proble, let's
now simply always the mouse cursor when pausing the emulation.