Update the profile test to use the int_test.tos binary instead of cyccheck.prg
cyccheck.prg is a bad idea for using it with "--mmu on", since the MMU
should get set up for accessing the IO registers properly - which we
don't do in our fake TOS that we use for the tests. The test was only
succeeding by accident so far, but it took a long time and any further
little change to cyccheck.prg seems to break it completely. Thus switch
the profile test to use the int_test.tos instead which does not depend
on proper MMU setup for its execution.